simplify readiness and healthcheck
the start_period has to be increased to 1hour for this change.
if the container is up and healty in less than that time, the probes
will be activated.

this will allow us to use docker-compose --wait and the container status
will be the same for alb and docker.

executing the searchd as main process is working fine with the trap, the
volume is cleaned up with:
* docker stop/kill
* pkill searchd inside container
* docker-compose down