fix issue with address rotation
without the dot in the -name parameter the address_metaphone* indexes
have been synced and rotated too.
since this index will be created right after the address index, it has
been synced in an instable state into the docker volume. this has
led to problems with the index rotation -> SegFault crash of searchd