this will fix an issue that has ocurred during the initial start of the
service container in the infra vhost environment. After a scan for
orphaned indexes the container (running in service mode) tried to remove
orphaned indexes on EFS which as mounted read-only on the host.

in service mode (as backend for search wsgi)
the service-sphinxsearch container needs read-only access to the EFS

in maintenance mode (triggered by data deploy on geodatasync)
the service-sphinxsearch container needs read-write access to the EFS

with this change the clean-up of orphaned indexes in the EFS will be
done only in maintenance mode, during the sphinx index createion.

this will allow us to strictly mount the geodata efs in read-only mode on all
our kubernetes or infra vhosts instances.