PB-82: a few changes in the dockerfile and in the efs sync script are
needed during the parallel operations of service-search-sphinx on
infra-vhost and k8s.

on k8s we are forced to mount the efs indexes with the full path:
so this will be the folder with the efs index files:
* /var/local/geodata/service-sphinxsearch/${DBSTAGING}/index
this is due to a limitation of the aws efs csi driver which is not
supporting subpaths as volumemount config. we would have to create new
peristant volumes with terraform which is unreasonable for this special
use-case.