improve readiness check for searchd and index freshness
some improvements in the sphinxsearch readiness probe:

- Ensure `searchd` is running before proceeding.
- Verify the existence and freshness of the index sync status file.
- Introduce `MAX_AGE` env variable with a default value of 300 seconds
to determine the maximum allowable age for the index sync status file.
- Prevent further tests if `index-sync-rotate.sh` is currently running
by checking for the presence of a lock file.

the check for the precence of a lock file will prevent false positives
in the future. some sync commands can take up to 10-15 minutes if bigger
indexes have been updated in the source EFS. during that time the
readiness probe has been reporting false positive alerts before.